% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map of Accidents}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer that corresponds to a state.}

\item{year}{An integer that corresponds to a year.}
}
\value{
This function plots a map of the locations of crashes for the chosen state in the chosen year.
}
\description{
This function prints maps of US states with dots representing the location of accidents for a given state in a given year. The user provides the state and year that will be mapped.
}
