% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read File}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string that corresponds to the name of the file to be opened.}
}
\value{
The function returns the contents of the filename in the form of a data frame. If the filename does not exist, the user will be notified.
}
\description{
This function is used to read in accident data files.
}
