% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Get data for Years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{An integer or string or list of integers and strings that correspond to years.}
}
\value{
This function returns a list of data frames that have a column for month and year of each accident. If there is a problem reading in or using the data an error will be returned.
}
\description{
This function is used to select the relevant columns of data from years specified by the user. This data is used by the function fars_summarize_years().
}
