% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize Accident Data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A integer or string or a list of integers or strings that are years.}
}
\value{
The function returns a tibble. There is a column for month, and for each year provided there is a column for the number of accidents in the US for that month.
}
\description{
This is a function that summarizes accident data by totaling the number of accidents in the country by month.
}
